Arrange codes, comments and files
