+++ QuickStart/AppDelegate.swift
@@ -9,7 +9,6 @@ import UIKit
 import CallKit
 import PushKit
 import SendBirdCalls
-import AVFoundation
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -24,49 +23,18 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // SendBirdCall.configure(appId: YOUR_APP_ID)
 
         self.autoSignIn { error in
-            guard let error = error else { return }
-            print(error.localizedDescription)
+            if error == nil { return }
+            // Show SignIn controller when failed to auto sign in
             self.window?.rootViewController?.present(UIStoryboard.signController(), animated: true, completion: nil)
         }
         
         // To process incoming call, you need to add `SendBirdCallDelegate` and implement its protocol methods.
         SendBirdCall.addDelegate(self, identifier: "com.sendbird.calls.quickstart.delegate")
         
+        // To receive incoming call, you need to register VoIP push token
         self.voipRegistration()
         
         return true
-    }
-    
-    func autoSignIn(completionHandler: @escaping (Error?) -> Void) {
-        // fetch credential
-        guard let pendingCredential = UserDefaults.standard.credential else {
-            DispatchQueue.main.async {
-                completionHandler(CredentialErrors.empty)
-            }
-            return
-        }
-        // authenticate
-        self.authenticate(with: pendingCredential, completionHandler: completionHandler)
-    }
-    
-    func authenticate(with credential: Credential, completionHandler: @escaping (Error?) -> Void) {
-        SendBirdCall.configure(appId: credential.appID)
-        let authParams = AuthenticateParams(userId: credential.userID, accessToken: credential.accessToken)
-        SendBirdCall.authenticate(with: authParams) { (user, error) in
-            guard user != nil else {
-                DispatchQueue.main.async {
-                    completionHandler(error ?? CredentialErrors.unknown)
-                }
-                return
-            }
-            let credential = Credential(accessToken: credential.accessToken)
-            let credentialManager = CredentialManager.shared
-            credentialManager.updateCredential(credential)
-            
-            DispatchQueue.main.async {
-                completionHandler(nil)
-            }
-        }
     }
     
     func applicationWillTerminate(_ application: UIApplication) {
@@ -91,51 +59,42 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // There might be some calls failed to be ended
         // In this case, I recommend that you register local notification to notify the unterminated calls.
     }
-}
-
-extension AppDelegate: PKPushRegistryDelegate {
-    func voipRegistration() {
-        self.voipRegistry = PKPushRegistry(queue: DispatchQueue.main)
-        self.voipRegistry?.delegate = self
-        self.voipRegistry?.desiredPushTypes = [.voIP]
-    }
     
-    func pushRegistry(_ registry: PKPushRegistry, didInvalidatePushTokenFor type: PKPushType) {
-        //
-    }
-    
-    // MARK: SendBirdCalls - Registering push token.
-    func pushRegistry(_ registry: PKPushRegistry, didUpdate pushCredentials: PKPushCredentials, for type: PKPushType) {
-        UserDefaults.standard.voipPushToken = pushCredentials.token
-        print("Push token is \(pushCredentials.token.toHexString())")
-        
-        SendBirdCall.registerVoIPPush(token: pushCredentials.token, unique: true) { error in
-            guard error == nil else { return }
+    // MARK: - SendBirdCall - Auto Sign In
+    func autoSignIn(completionHandler: @escaping (Error?) -> Void) {
+        // fetch credential
+        guard let pendingCredential = UserDefaults.standard.credential else {
+            DispatchQueue.main.async {
+                completionHandler(CredentialErrors.empty)
+            }
+            return
         }
+        // authenticate
+        self.authenticate(with: pendingCredential, completionHandler: completionHandler)
     }
     
-    // MARK: SendBirdCalls - Receive incoming push event
-    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType) {
-        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type, completionHandler: nil)
-    }
-    
-    // Handle Incoming pushes
-    func pushRegistry(_ registry: PKPushRegistry, didReceiveIncomingPushWith payload: PKPushPayload, for type: PKPushType, completion: @escaping () -> Void) {
-        // MARK: Handling incoming call
-        SendBirdCall.pushRegistry(registry, didReceiveIncomingPushWith: payload, for: type) { uuid in
-            guard uuid != nil else {
-                let update = CXCallUpdate()
-                update.remoteHandle = CXHandle(type: .generic, value: "invalid")
-                let randomUUID = UUID()
-
-                CXCallManager.shared.reportIncomingCall(with: randomUUID, update: update) { _ in
-                    CXCallManager.shared.endCall(for: randomUUID, endedAt: Date(), reason: .acceptFailed)
+    func authenticate(with credential: Credential, completionHandler: @escaping (Error?) -> Void) {
+        // Configure app ID before authenticate
+        SendBirdCall.configure(appId: credential.appID)
+        
+        // Authenticate
+        let authParams = AuthenticateParams(userId: credential.userID, accessToken: credential.accessToken)
+        SendBirdCall.authenticate(with: authParams) { (user, error) in
+            guard user != nil else {
+                // Failed
+                DispatchQueue.main.async {
+                    completionHandler(error ?? CredentialErrors.unknown)
                 }
-                completion()
                 return
             }
-
-            completion()
+            // Succeed
+            let credential = Credential(accessToken: credential.accessToken)
+            let credentialManager = CredentialManager.shared
+            credentialManager.updateCredential(credential)
+            
+            DispatchQueue.main.async {
+                completionHandler(nil)
+            }
         }
     }
 }
