+++ QuickStart/View/Base.lproj/Main.storyboard
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14868" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="vv4-4l-vzD">
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="16097.3" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="vv4-4l-vzD">
     <device id="retina6_1" orientation="portrait" appearance="light"/>
     <dependencies>
         <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14824"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="16087"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
     </dependencies>
@@ -21,7 +21,7 @@
         <!--Voice Call View Controller-->
         <scene sceneID="xQ3-Jv-7ju">
             <objects>
-                <viewController storyboardIdentifier="VoiceCallViewController" id="7ra-kc-2Kq" customClass="VoiceCallViewController" customModule="SendBird_Calls" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="VoiceCallViewController" id="7ra-kc-2Kq" customClass="VoiceCallViewController" customModule="Sendbird_Calls" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="eVj-7G-Q5q">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -145,7 +145,7 @@
         <!--Dial-->
         <scene sceneID="eda-BC-zTo">
             <objects>
-                <viewController storyboardIdentifier="DialViewController" id="QH3-LJ-tLk" customClass="DialViewController" customModule="SendBird_Calls" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="DialViewController" id="QH3-LJ-tLk" customClass="DialViewController" customModule="Sendbird_Calls" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="rcF-ir-O6k">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
@@ -283,7 +283,6 @@
                             <constraint firstItem="ZIt-Bt-jdD" firstAttribute="leading" secondItem="Ue9-qz-Py8" secondAttribute="trailing" constant="8" id="Twc-8X-p4Z"/>
                             <constraint firstItem="Jpp-zS-JFC" firstAttribute="trailing" secondItem="Eg3-1U-GuL" secondAttribute="trailing" id="Zjy-ZQ-pCQ"/>
                             <constraint firstItem="QG9-oG-JhN" firstAttribute="leading" secondItem="Jpp-zS-JFC" secondAttribute="leading" constant="24" id="csT-SZ-fsd"/>
-                            <constraint firstItem="Eg3-1U-GuL" firstAttribute="top" secondItem="q8c-GM-NKq" secondAttribute="bottom" constant="4.5" id="hc5-SM-HYU"/>
                             <constraint firstItem="Ue9-qz-Py8" firstAttribute="top" secondItem="Jpp-zS-JFC" secondAttribute="top" constant="5" id="knw-Sr-crq"/>
                             <constraint firstItem="Jpp-zS-JFC" firstAttribute="trailing" secondItem="LnJ-3y-Xyu" secondAttribute="trailing" constant="24" id="mff-Li-ncj"/>
                             <constraint firstItem="twf-Fk-qMH" firstAttribute="centerX" secondItem="Jpp-zS-JFC" secondAttribute="centerX" constant="-48" id="neL-hJ-mc7"/>
@@ -292,6 +291,7 @@
                             <constraint firstItem="LnJ-3y-Xyu" firstAttribute="leading" secondItem="Jpp-zS-JFC" secondAttribute="leading" constant="24" id="rG4-8e-wfv"/>
                             <constraint firstItem="Ue9-qz-Py8" firstAttribute="leading" secondItem="Jpp-zS-JFC" secondAttribute="leading" constant="24" id="tLo-ms-kSH"/>
                             <constraint firstItem="Jpp-zS-JFC" firstAttribute="trailing" secondItem="q8c-GM-NKq" secondAttribute="trailing" constant="24" id="xWW-6f-lf8"/>
+                            <constraint firstItem="Eg3-1U-GuL" firstAttribute="top" secondItem="Ue9-qz-Py8" secondAttribute="bottom" constant="7" id="y8E-b7-rof"/>
                             <constraint firstItem="ZIt-Bt-jdD" firstAttribute="top" secondItem="Ue9-qz-Py8" secondAttribute="top" id="zRt-54-zs7"/>
                         </constraints>
                         <viewLayoutGuide key="safeArea" id="Jpp-zS-JFC"/>
@@ -360,7 +360,7 @@
         <!--Video Call View Controller-->
         <scene sceneID="95y-fH-Uoo">
             <objects>
-                <viewController storyboardIdentifier="VideoCallViewController" id="Suw-Gg-jyg" customClass="VideoCallViewController" customModule="SendBird_Calls" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="VideoCallViewController" id="Suw-Gg-jyg" customClass="VideoCallViewController" customModule="Sendbird_Calls" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="wEz-Ud-OC4">
                         <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
